websocket/requirements.txt
django==3.1.7
mysqlclient==2.0.3
requests==2.25.1
channels==3.0.3
channels-redis==3.3.0
djangorestframework==3.12.4